refactor/lint: #299 modelos db (ver mensaje del commit -->)
\## Cambios en src/app/@core/data/models/:

Habían modelos repetidos en algunas partes y/o que, vía herencia;
se pueden dividir en partes menores, sobretodo para bajar la carga
de las transacciones y a su vez debería de impactar de manera
positiva en el rendimiento del cliente.

Los archivos que se eliminaron, se eliminaron principalmente porque
ya habían otros modelos existentes, por ejemplo para el catalogo y
subgrupo1/2 y otros, como tal se movieron a archivos dentro de
la carpeta de modelos asociados al catálogo

Lo que principalmente se modificó fue todo lo asociado a los modelos
de catálogos, grupos y subgrupos, moviendo los elementos de grupos y
subgrupos a un nuevo archivo jerarquia.ts; y las transacciones al
otro nuevo archivo transacciones.ts.

Los demás cambios de esta carpeta obedecen principalmente a lo anterior.

\## Cambios en los demás lados:

Obedecen a propagar los anteriores cambios. En los casos en que se
eliminaron archivos (dentro de src/app/@core/data/models/), lo único
que se hizo fue actualizar los `import` conforme a todo lo anterior.
