--- src/app/@core/data/models/catalogo.ts
@@ -1,9 +1,0 @@
-
-export class Catalogo {
-  Id: number;
-  Nombre: string;
-  Descripcion: string;
-  FechaInicio: Date;
-  FechaFin: Date;
-  Activo: boolean;
-}
