+++ src/app/@core/data/models/catalogo/catalogo.ts
@@ -1,5 +1,8 @@
-export class Catalogo {
+export class CatalogoID {
     Id: number;
+}
+
+export class Catalogo extends CatalogoID {
     Nombre: string;
     Descripcion: string;
     FechaInicio: Date;
