--- src/app/@core/data/models/grupo.ts
@@ -1,9 +1,0 @@
-import { Catalogo } from './catalogo';
-
-export class Grupo {
-  Id: number;
-  Nombre: string;
-  Descripcion: string;
-  Activo: boolean;
-  Catalogo: Catalogo;
-}
