--- src/app/@core/data/models/subgrupo_1.ts
@@ -1,9 +1,0 @@
-import { Grupo } from './grupo';
-
-export class Subgrupo1 {
-  Id: number;
-  Nombre: string;
-  Descripcion: string;
-  Activo: boolean;
-  Grupo: Grupo;
-}
