--- src/app/@core/data/models/subgrupo_2.ts
@@ -1,9 +1,0 @@
-import { Subgrupo1 } from './subgrupo_1';
-
-export class Subgrupo2 {
-  Id: number;
-  Nombre: string;
-  Descripcion: string;
-  Activo: boolean;
-  Subgrupo1: Subgrupo1;
-}
