+++ src/app/pages/catalogo-bienes/registro-cuentas-catalogo/registro-cuentas-catalogo.component.ts
@@ -1,9 +1,8 @@
 import { Component, OnInit, Input, Output, EventEmitter, ViewChild, QueryList, ViewChildren } from '@angular/core';
-import { Subgrupo } from '../../../@core/data/models/catalogo/subgrupo';
 import { TranslateService, LangChangeEvent } from '@ngx-translate/core';
 import { CatalogoElementosHelper } from '../../../helpers/catalogo-elementos/catalogoElementosHelper';
-import { Grupo } from '../../../@core/data/models/catalogo/grupo';
-import { Catalogo } from '../../../@core/data/models/catalogo';
+import { Grupo, Subgrupo } from '../../../@core/data/models/catalogo/jerarquia';
+import { Catalogo } from '../../../@core/data/models/catalogo/catalogo';
 import Swal from 'sweetalert2';
 import { Store } from '@ngrx/store';
 import { IAppState } from '../../../@core/store/app.state';
