+++ src/app/pages/catalogo/crud-catalogo/crud-catalogo.component.ts
@@ -1,5 +1,5 @@
 
-import { Catalogo } from '../../../@core/data/models/catalogo';
+import { Catalogo } from '../../../@core/data/models/catalogo/catalogo';
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { FORM_CATALOGO } from './form-catalogo';
 import { ToasterService, ToasterConfig, Toast, BodyOutputType } from 'angular2-toaster';
